requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-jet-reboot==1.3.7
django-multiselectfield==0.1.12
django-stubs==4.2.7
django-stubs-ext==4.2.7
django-types==0.19.1
djangorestframework==3.14.0
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sqlparse==0.4.4
types-psycopg2==2.9.21.20240201
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
